+++ Changes.md
@@ -1,3 +1,5 @@
+Release 19.1.0
+==============
 * #1129 - Display scores for other judges in the subjective score summary
 * shrink font size for judging group on pit signs
 * #1215 - Allow the user to specify the order of the tables in performance schedules